+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.23.0]
+
+### Added
+
+- Android and iOS/tvOS: Expose `VideoQuality.colorInfo.dynamicRange` (`sdr`, `hdr`, `unknown`) via the new `DynamicRange` enum
+
+### Changed
+
+- Update Bitmovin's native iOS SDK version to [`3.118.0`](https://developer.bitmovin.com/playback/docs/release-notes-ios#31180)
+- Update Bitmovin's native Android SDK version to [`3.160.0+jason`](https://developer.bitmovin.com/playback/docs/release-notes-android#31600)
+
 ## [1.22.0] - 2026-07-23
 
 ### Added